Optimize for smaller phones (iPhone 6/7/8/SE2)
Squeeze UI elements a bit closer together so they do not overlap on
smaller phones. (The GUI is probably still too big for iPhone 5 screen
resolutions but practically no phones are being sold today that are that
small.) Also disable scrolling and selecting text as they just get in
the way.